Add TreeModel base class and provide generic display implementation for it
